# swagger.yml
openapi: 3.0.0
info:
  title: Task Manager
  version: 3.1.6
servers:
  - url: http://localhost:8080/
    description: Dev server
paths:
...
  /projects:
    post:
      summary: Create Project
      tags:
        - Projects
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectToCreate"
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        '400':
          description: bad request
        '500':
          description: internal server error
    get:
      summary: User projects
      tags:
        - Projects
      operationId: getUserProjects
      responses:
        '200':
          description: Successful response with project received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Projects"
        '404':
          description: not found
        '500':
          description: internal server error
...
components:
  schemas:
...
    Project:
      type: object
      required:
        - id
        - name
        - user_id
        - created_at
      properties:
        id:
          type: integer
          example: 15
        name:
          type: string
          example: Project X
        user_id:
          type: integer
          example: 4
        created_at:
          type: string
          format: 2024-05-15
    Projects:
      type: array
      items:
        $ref: "#/components/schemas/Project"
    ProjectToCreate:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          example: 15
...